feature/login and routes permissions

+++ src/controllers/user.controllers.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from "express";
+import { AppError } from "../errors";
 import { iUser, iUserReturn, iUsersReturn } from "../interfaces/users.interfaces";
-import createUserService from "../services/user/createUser.services";
+import createUserService from "../services/user/createUser.service";
 import deleteUserService from "../services/user/deleteUser.service";
 import updateUserService from "../services/user/patchUser.service";
 import readUsersService from "../services/user/readUsers.service";
@@ -16,6 +17,7 @@ const createUserController = async (req: Request, res: Response) => {
 }
 
 const readUsersController = async (req: Request, res: Response) => {
+    if(!req.user.admin) throw new AppError("Insufficient Permission", 403);
 
     const users: iUsersReturn = await readUsersService()
 
@@ -24,6 +26,8 @@ const readUsersController = async (req: Request, res: Response) => {
 }
 
 const deleteUserController = async (req: Request, res: Response) => {
+    const userId = Number(req.params.id)
+    if(!req.user.admin && req.user.id !== userId) throw new AppError("Insufficient Permission", 403);
 
     await deleteUserService(Number(req.params.id))
 
@@ -32,9 +36,10 @@ const deleteUserController = async (req: Request, res: Response) => {
 }
 
 const updateUserController = async (req: Request, res: Response) => {
-
-    const userData = req.body
     const userId = Number(req.params.id)
+    const userData = req.body
+    if(!req.user.admin && req.user.id !== userId) throw new AppError("Insufficient Permission", 403);
+
 
     const updatedUser = await updateUserService(userData, userId)
